Fix Sudo Log Ownership for non-'jenkins' users
If builds are run with a user other than 'jenkins', chwon'ing of the
sudoers log will fail. Switching the command to use the current
user (and user's default login group) fixes this.

Note: See chwon(1) DESCRIPTION for an explination of using a colon after
the username while omitting the group

Change-Id: Ia9b96e93a250fd22eb36c94471a06c2e211dc9e3
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>