Feat: Choose best python version to run venv
We've seen problems with capture-instance-metadata.sh choosing the
best python path, so this should help get the best version on each
builder.

If the $PYTHON variable is set, this will be used. If not, we check to
see if python3 is available, as this should point to the latest
version. If this is also not available, we run with the basic python
command.

Change-Id: I9950cc286c72fd17eac7a3c678ef8ca04ccd8921
Signed-off-by: Eric Ball <eball@linuxfoundation.org>