Fix: safer behaviour handling java variables
Detect both Debian and Ubuntu when configuring java runtime to work around OS detection bugs. This applies to certain Docker containers that may not have the lsb-release package installed. Improved the handling of an unset variable, where the java runtime setup would get broken when it was previously working during an earlier run of the script.

Change-Id: I69c0756d740430dabef823fa8eed8cbf67e343c6
Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>