Fix: JAVA_HOME directory detection
OpenDaylight jenkins maven jobs with jdk17 and CentOS8 currently fails
because the JAVA_HOME variable is not correctly set.

https://jenkins.opendaylight.org/releng/job/transportpce-maven-verify-chlorine-mvn38-openjdk17/7/console

It appears that the architecture is now appended to the folder name
and that Debian and CentOS GNU/Linux distributions use at that time
different conventions names (amd64 and x86_64 respectively).

- add a folder exisitence check in related script
- if no folder found try to find an approaching solution
  using ls -d and a * blob appended to the folder basename

Signed-off-by: guillaume.lambert <guillaume.lambert@orange.com>
Change-Id: I585cb34e8126ac5827ae33b5c1ed771fd78b3d10