Fix: Check for $lf_env in lf-activate-venv()
Check for the presense of an existing venv ("/tmp/.os_lf_venv")
into the lf-activate-venv() for the openstack dependencies which
makes the code look cleaner to maintain.

The file "/tmp/.os_lf_venv" is only created by python-tools-install.sh
which is called during pre-build.

If the dependency is already installed pip will skip installing the
package.

Issue-ID: RELENG-4403
Change-Id: I559f759a8dba7eca0a62f8b73a360dc627699ed2
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>