Fix: Copy the spdx file into the version dir
The SBOM generator script creates an spdx file in the root level of m2repo
of the $project. When the artifacts are pushed the spdx file gets
overwritten.

Find all the artifact version directories that match the release version
within the namespace and copy the spdx file into each of the version
directories.

Change-Id: Ia8bd06ac160e30886c7133aef8f0c82e5aded3dd
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>