WIP Fix: Merge bundle after checking detached head
Address the problem where the tag is not merged into master
i.e. the release is not part of the master history.

Check commit count between the HEAD and origin/${GERRIT_BRANCH}
before the fetch and merge operation. This is done to ensure the
git bundle along with the tag object is pushed on the correct
branch.

Change-Id: I675c815d4ac39244adf5168c1fbd051f8c626290
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>