Upgrade!: JJB code to comply with v5.x
There has been an update to the YAML parser with more fine-tuned
control over YAML parsing. YAML objects and parameter expansion logic
is in accordance with the latest YAML specification 1.2.2 (released
in Oct 2021).

Notable changes include:
- Aggressive parameter expansion. This may lead to parameters expanded
in places where they were not expanded before.
- Recursive parameters usage is being treated as error.
- Strict checking for missing parameters.
- Files included using '!include-raw:' elements and having formatting
  it path ('lazy-loaded' in previous implementation) are now expanded too.
  Use '!include-raw-escape:' for them instead.

NOTE: this is a beaking change would have potential impact
many of jobs in global-jjb and in ci-man repo that DO NOT use
our templates.

Issue-ID: RELENG-4686
Change-Id: Ie7417956bdcf5f4d6477dbff87baca6b1b8aeeaf
Ref: https://groups.google.com/u/2/g/jenkins-job-builder/c/1fx7PmADgXw
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>