Fix: update submodules in jjb-deploy-job
The jjb-deploy-job can be used to test upgrades
in ci-management related repositories.
Those contain submodules, e.g. global-jjb.

The parent way to checkout the change under test
does not update the submodules,
so the job uses a combination of unmerged main repo
with surrent master submodules.
That is unhelpful and dangerous.

This change adds a command
to synchronize submodules after checkout.

Change-Id: I407ccf86e404bd01675439bb4df8c85a804267bc
Signed-off-by: Vratko Polak <vrpolak@cisco.com>