CI: Switch to using reuse for license checking
The tox based license checker is failing due to some issues related to
distutils. As we do not really maintain the validator and there is a
better upstream system we're going to switch to it.

This change brings in all the needed modifications to switch tools and
also take care of the license and copyright information without having
to touch all of the files in the repository.

Issue: RELENG-5356
Change-Id: Ibaecb20f922adf5e7a050be9f03f945076dd2dfe
Ref: https://github.com/fsfe/reuse-tool
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>