Chore: Add v3-*-sjc flavors to openstack script
v3 flavors in Vexxhost have different hardware IDs in yul and sjc,
presenting an issue we didn't have with older flavors. This will allow
projects in sjc to specify the flavor with "-sjc" at the end in order
to correctly resolve the hardware ID.

For the sake of simplicity, only the most common flavors have been
added.

Change-Id: Ifff7a0ce45fb253e653340661c3d7482f168a98b
Signed-off-by: Eric Ball <eball@linuxfoundation.org>